backend/requirements.txt
Django>=5.2.9
djangorestframework
django-cors-headers
django-ckeditor
Pillow
django-jazzmin
